+++ test/server/fetch_secrets.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+# staging environment
+op item get gkrzrqajgcmg4mwkjnbkdsndsi --fields notesPlain | tr -d '\"' > .env.staging
+
+# prod environment
+op item get 2g5wv7lcfkmvwqfpjid2uf4v5a --fields notesPlain | tr -d '\"' > .env.prod