+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add raw JWT to Go context when using the scoped permissions JWT validators
 - Updated the flow for detecting if JWT middleware should be used 
 - Updated execution inputs to use `Global.user` derived from Superblocks JWT
+- Updated claims required to be present in a JWT for agent authorization